🚧 Update parse and filter

+++ src/assets/assets.controller.ts
@@ -3,6 +3,9 @@ import {
   Controller,
   Delete,
   Get,
+  HttpCode,
+  HttpException,
+  HttpStatus,
   Param,
   Patch,
   Post,
@@ -15,6 +18,7 @@ import { AssetsService } from './assets.service';
 import { CreateAssetDto } from './dto/create-asset.dto';
 import { UpdateAssetDto } from './dto/update-asset.dto';
 import { Asset } from './interfaces/asset.interface';
+import { ParseObjectIdPipe } from './pipe/parse-objectid.pipe';
 
 @Controller('assets')
 export class AssetsController {
@@ -26,8 +30,12 @@ export class AssetsController {
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string): Promise<Asset> {
-    return this.assetsService.findOne(id);
+  async findOne(@Param('id', ParseObjectIdPipe) id: string): Promise<Asset> {
+    const asset = await this.assetsService.findOne(id);
+    if (!asset) {
+      throw new HttpException('Asset not found', HttpStatus.NOT_FOUND);
+    }
+    return asset;
   }
 
   @Post()
@@ -39,13 +47,25 @@ export class AssetsController {
 
   @Patch(':id')
   @UseGuards(AuthGuard('basic'))
-  update(@Param('id') id: string, @Body() updateAssetDto: UpdateAssetDto) {
+  async update(
+    @Param('id', ParseObjectIdPipe) id: string,
+    @Body() updateAssetDto: UpdateAssetDto,
+  ) {
+    const asset = await this.assetsService.findOne(id);
+    if (!asset) {
+      throw new HttpException('Asset not found', HttpStatus.NOT_FOUND);
+    }
     return this.assetsService.update(id, updateAssetDto);
   }
 
   @Delete(':id')
   @UseGuards(AuthGuard('basic'))
-  remove(@Param('id') id: string): Promise<Asset> {
+  @HttpCode(HttpStatus.NO_CONTENT)
+  async remove(@Param('id', ParseObjectIdPipe) id: string): Promise<Asset> {
+    const asset = await this.assetsService.findOne(id);
+    if (!asset) {
+      throw new HttpException('Asset not found', HttpStatus.NOT_FOUND);
+    }
     return this.assetsService.remove(id);
   }
 }
